Se corrigen errores de código y de documentación
