+++ src/swagger.js
@@ -21,7 +21,10 @@ const swaggerDefinition = {
 
 const options = {
   swaggerDefinition,
-  apis: [`${__dirname}/components/user/adapters/routes/user.routes.js`],
+  apis: [
+    `${__dirname}/components/user/adapters/routes/user.routes.js`,
+    `${__dirname}/components/processOnReturn/adapters/routes/processOnReturn.routes.js`,
+  ],
 };
 
 const swaggerSpec = swaggerJSDoc(options);
